refactor: split GlobalExceptionHandler into per-modules exception handlers (#128)
